Add launchapp key

+++ android/app/src/main/java/com/reactkeyboard/ime/MyInputMethodService.java
@@ -1,13 +1,16 @@
 package com.reactkeyboard.ime;
 
+import android.content.Intent;
 import android.inputmethodservice.InputMethodService;
 import android.inputmethodservice.Keyboard;
 import android.inputmethodservice.KeyboardView;
+import android.net.Uri;
 import android.text.TextUtils;
 import android.view.View;
 import android.view.inputmethod.InputConnection;
 
 import com.reactkeyboard.R;
+import com.reactkeyboard.ime.constaint.AppConstants;
 
 public class MyInputMethodService extends InputMethodService implements KeyboardView.OnKeyboardActionListener {
 
@@ -31,25 +34,37 @@ public class MyInputMethodService extends InputMethodService implements Keyboard
     }
 
     @Override
-    public void onKey(int primatyCode, int[] keyCodes) {
+    public void onKey(int primaryCode, int[] keyCodes) {
         InputConnection inputConnection = getCurrentInputConnection();
 
         if (inputConnection != null) {
-            switch(primatyCode) {
-                case Keyboard.KEYCODE_DELETE :
-                    CharSequence selectedText = inputConnection.getSelectedText(0);
+            if(String.valueOf(primaryCode).startsWith(String.valueOf(AppConstants.MENU_PREFIX_CODE)) ) {
+                switch (primaryCode){
+                    case AppConstants.MENU1:
+                        inputConnection.commitText("Hello world!", 1);
+                        break;
+                    case AppConstants.MENU2:
+                        openApp();
+                        break;
+                }
+            } else {
+                switch(primaryCode) {
+                    case Keyboard.KEYCODE_DELETE :
+                        CharSequence selectedText = inputConnection.getSelectedText(0);
 
-                    if (TextUtils.isEmpty(selectedText)) {
-                        inputConnection.deleteSurroundingText(1, 0);
-                    } else {
-                        inputConnection.commitText("", 1);
-                    }
+                        if (TextUtils.isEmpty(selectedText)) {
+                            inputConnection.deleteSurroundingText(1, 0);
+                        } else {
+                            inputConnection.commitText("", 1);
+                        }
 
-                    break;
-                default :
-                    char code = (char) primatyCode;
-                    inputConnection.commitText(String.valueOf(code), 1);
+                        break;
+                    default :
+                        char code = (char) primaryCode;
+                        inputConnection.commitText(String.valueOf(code), 1);
+                }
             }
+
         }
     }
 
@@ -76,5 +91,19 @@ public class MyInputMethodService extends InputMethodService implements Keyboard
     @Override
     public void swipeUp() {
 
+    }
+
+    private void openApp(){
+        Intent launchIntent = null;
+
+        try{
+            launchIntent = getPackageManager().getLaunchIntentForPackage(AppConstants.applicationId);
+        } catch (Exception ignored) {}
+
+        if(launchIntent == null){
+            startActivity(new Intent(Intent.ACTION_VIEW).setData(Uri.parse("https://play.google.com/store/apps/details?id=" + AppConstants.applicationId)));
+        } else {
+            startActivity(launchIntent);
+        }
     }
 }
